requirements.txt
annotated-types==0.6.0
anyio==4.3.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.28.0
